news-verification-project/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
pyjwt==2.8.0
werkzeug==2.3.7
python-dotenv==1.0.0
tensorflow==2.13.0
opencv-python==4.8.1.78
scikit-learn==1.3.1
numpy==1.24.3
pillow==10.0.1
pyspark==3.4.1
pandas==2.1.1 